+++ tardis/montecarlo/montecarlo_numba/utils.py
@@ -1,0 +1,16 @@
+import numpy as np
+from numba import njit
+from tardis.montecarlo.montecarlo_numba import (
+    njit_dict,
+    numba_config,
+    njit_dict_no_parallel,
+)
+
+
+class MonteCarloException(ValueError):
+    pass
+
+
+@njit(**njit_dict_no_parallel)
+def get_random_mu():
+    return 2.0 * np.random.random() - 1.0
