+++ tardis/montecarlo/montecarlo_numba/vpacket.py
@@ -2,7 +2,7 @@ from numba import float64, int64, boolean
 from numba import njit, gdb
 from numba.experimental import jitclass
 
-from tardis.montecarlo.montecarlo_numba import njit_dict , njit_dict_no_parallel
+from tardis.montecarlo.montecarlo_numba import njit_dict, njit_dict_no_parallel
 from tardis.montecarlo import (
     montecarlo_configuration as montecarlo_configuration,
 )
@@ -10,16 +10,23 @@ import math
 import numpy as np
 
 from tardis.montecarlo.montecarlo_numba.r_packet import (
-    calculate_distance_boundary,
-    get_doppler_factor,
-    calculate_distance_line,
-    calculate_tau_electron,
     PacketStatus,
     move_packet_across_shell_boundary,
-    angle_aberration_LF_to_CMF,
-    angle_aberration_CMF_to_LF,
     test_for_close_line,
 )
+
+from tardis.montecarlo.montecarlo_numba.calculate_distances import (
+    calculate_distance_boundary,
+    calculate_distance_line,
+)
+
+from tardis.montecarlo.montecarlo_numba.frame_transformations import (
+    get_doppler_factor,
+    angle_aberration_LF_to_CMF,
+    angle_aberration_CMF_to_LF,
+)
+
+from tardis.montecarlo.montecarlo_numba.opacities import calculate_tau_electron
 
 vpacket_spec = [
     ("r", float64),
